Drop 32-bit platforms and WatchOS
This allows us to unify the implementation around 64-bit for now and avoid duplication.
